script mod

+++ daxby/simple-pim/Param.h
@@ -1,7 +1,8 @@
     #ifndef PARAM_H
     #define PARAM_H
+    #include <stdint.h>
     #include <stdlib.h>
-    typedef double T; 
+    typedef int32_t T; 
     const uint32_t dpu_number = 32;
     uint32_t print_info = 0;
     uint64_t nr_elements = 1048576;
